Moved generate_unique_slug from utils.py to helpers.py
